Feat: 입력을 주관하는 User를 생성하여 분리
