+++ src/main/java/baseball/game/BaseballApp.java
@@ -1,7 +1,5 @@
 package baseball.game;
 
-import camp.nextstep.edu.missionutils.Console;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -9,15 +7,16 @@ import java.util.Map;
 public class BaseballApp {
 
     public Computer computer;
+    public User user;
 
     public BaseballApp() {
-        this.computer = new Computer();
+        this.user = new User();
+        this.computer = new Computer(user);
     }
 
     public void run() {
 
         System.out.println("숫자 야구 게임을 시작합니다.");
-        computer.genNumbers();
         startGame();
 
     }
@@ -28,9 +27,9 @@ public class BaseballApp {
 
         while(power) {
 
-            String input = getInput();
+            String input = user.getInput();
 
-            if(checkInput(input)) {
+            if(user.checkInput(input)) {
                 Map<String, Integer> compareMap = computer.compare(transToList(input));
                 boolean result = computer.showResult(compareMap);
 
@@ -39,28 +38,6 @@ public class BaseballApp {
                 }
             }
         }
-    }
-
-    public String getInput() {
-
-        System.out.print("숫자를 입력해주세요 : ");
-
-        return Console.readLine();
-    }
-
-    public boolean checkInput(String input) {
-
-        if(input.length() != 3) {
-            throw new IllegalArgumentException("3자리의 숫자가 입력되어야 합니다.");
-        }
-
-        try {
-            Integer.parseInt(input);
-        } catch(NumberFormatException e) {
-            throw new IllegalArgumentException("숫자가 아닌 값이 입력되었습니다.");
-        }
-
-        return true;
     }
 
     public List<Integer> transToList (String input) {
